[Bug] Pressing on the space bar after moving to the next word
If you move to the next word and then you press on the "space" button. It will skip the current word and move on to the next, because although the model is hidden however the button does still have focus.
Had to truly hide the model by changing the tabIndex if showModal is false, and blur the button to remove focus